requirements.txt
# Flask dan server
flask==2.3.3
gunicorn==21.2.0

# Google Cloud
google-cloud-pubsub==2.18.4
google-cloud-aiplatform==1.36.4
google-cloud-secret-manager==2.16.4
google-cloud-logging==3.8.0
google-auth==2.23.0
google-api-python-client==2.108.0

# Utilitas
requests==2.31.0
python-dotenv==1.0.0
retry==0.9.2

# Email processing
email-validator==2.1.0
html2text==2020.1.16

# Untuk testing
pytest==7.4.3
pytest-mock==3.12.0
